// tutorialls/src/validation/zod/form/card/create.schema.ts
import { z } from 'zod';

export const createTutorialFormSchema = z.object({
  title: z.string().min(1, 'Title is required'),
  content: z.string().min(1, 'Content is required'),
  author: z.string().min(1, 'Author is required'),
});

export type ICreateTutorialForm = z.infer<typeof createTutorialFormSchema>;